fix(BaseballGameService) : 기능 분리
 - RandomBoxService에 있던 사용자 입력 기능 BaseballGameService로 이동
