+++ src/main/java/nextstep/baseball/service/BaseballGameService.java
@@ -1,0 +1,7 @@
+package nextstep.baseball.service;
+
+import java.util.List;
+
+public interface BaseballGameService {
+    List<Integer> inputUserNumber(String str);
+}
