+++ src/main/java/nextstep/baseball/service/RandomBoxService.java
@@ -3,7 +3,6 @@ package nextstep.baseball.service;
 import java.util.List;
 
 public interface RandomBoxService {
-    List<Integer> inputUserNumber(String str);
     void setRandomBox();
     List<Integer> createRandomNumber();
 }
